requirements.txt
pandas==2.0.3
numpy==1.24.3
seaborn==0.12.2
matplotlib==3.7.1
scikit-learn==1.2.2
ibm-watson==6.1.0
python-dotenv==1.0.0
flask
gunicorn
